[WNMGDS-976] Convert StepList component to TypeScript (#1453)
* Convert StepLink to TypeScript

* Convert Step and SubStep to TypeScript, but need TS 4.x features

* Don't need to explicitly define this type with TS 4.5

* Convert StepList to TypeScript

* Convert StepList index file and export types

* Fix StepList story not showing up in Storybook

* onClick handler has never been treated as required, so don't require it

* Don't require substeps either!

* Missed this when updating onClick requirements

* Cleaned out old typedef files
