--- packages/design-system/src/components/StepList/index.js
@@ -1,1 +1,0 @@
-export { default as StepList } from './StepList';
