+++ packages/design-system/src/components/StepList/index.ts
@@ -1,0 +1,2 @@
+export * from './StepList';
+export type { StepObject } from './Step';
