+++ packages/design-system/src/types/index.d.ts
@@ -4,7 +4,6 @@
 export { default as Dialog } from './Dialog/Dialog';
 export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
 export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
-export { default as StepList } from './StepList/StepList';
 
 // Type definitions generated by cmsds scripts
 export * from './Accordion';
@@ -24,6 +23,7 @@ export * from './Pagination';
 export * from './Review';
 export * from './SkipNav';
 export * from './Spinner';
+export * from './StepList';
 export * from './Table';
 export * from './Tabs';
 export * from './TextField';
